[ADD & MODIFY] 1. add fcn for segmentation. 2. use absolute path in import instead of relative path. 3. fix bugs in base_dataset.py(compare output shapes and a small bug). 4. modify labels dataset dtype to tf.int32. 5. add loggings in test case.

+++ tensorbob/dataset/dataset_utils.py
@@ -236,7 +236,7 @@ def _get_classification_labels_dataset(dataset_config):
     :return:
     """
     src = dataset_config['src']
-    return tf.data.Dataset.from_tensor_slices(src)
+    return tf.data.Dataset.from_tensor_slices(tf.to_int32(src))
 
 
 ###################################### 获取 图像分割 标签 ##################################################
